Nao lola testing on aragorn (#1)
Fixed some issues faced when testing on a real Nao.
Care had to be taken when unpacking msgpack messages, as they may contains null characters. If read into a string, this will terminate the string and cause strange issues. This is why a std::array is used when passing around the data.